extract test Mission/Rocket creation to util classes for future tests
